Extract scrapin related code to separate service
